requirements.txt
mne
psycopg2-binary
redis